requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
convcolors==2.2.0
extcolors==1.0.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.5.0
Werkzeug==2.3.6
